Update noir version and setup
The first pass at this was before the noir template worked for lein2,
so we atempted to copy over what we could by hand.  It has been updated
and we know more about noir, so look at it for how to load views.  Also
move bootstrap around since noir serves the resources/ directory.